refactor: split handlers.ts (1327→203 lines) + fix empty catch blocks
- #12: Extract handleSearch, handleMap, handleMap3d, handleStats, handleLearn
  into focused modules (search.ts, map.ts, map3d.ts, stats.ts, learn.ts)
- #13: Add explanatory comments to all empty catch blocks across
  dashboard.ts, scripts, and vector adapters
- All functions now <50 lines, all files <800 lines
- Re-exports from handlers.ts preserve existing import paths
